Added new font generator project files
Still needs some fine tuning, so there will be some updates in the
coming days. Also need to finish rewriting the module for better memory
management and to support the new generator. Additionally I'd like to
rewrite the old python to support this format as well. It won't be as
good as the C# version, but then at least it's not completely left for
dead.